chore(): refactor scroll restoration functions
